feat: Add snapshot capability <INCOMPLETE>
This should be a feature of the repository interface rather than the
aggregate, since logic associated with converting aggregate instances to
and restoring them from serialized snapshots is not within the scope of
the domain logic.

As such, this should sit in the infrastructure layer, as an optional
argument when constructing the repository. Snapshot stores should be the
ones to define how they serialize / deserialize their information, not
the domain model objects themselves.